refactor: 精査結果に基づく設計是正
- group-modal-builder.jsをcalendar-group-managerに統合（唯一の利用元で密結合）
- CalendarGroupManagerの動的importをモジュールレベルimportに修正
- ReminderService(61行)を削除しコントローラーにインライン化（薄すぎるラッパー）
- AuthServiceを削除しコントローラーにDOM操作を戻す（DOM-freeパターン違反）
- event-handlers.jsのregisterEvent重複3箇所を_registerGoogleEvent/_registerLocalEventに統一
- 残課題をTODO.mdのリファクタリング（設計改善）セクションに記載
